requirements.txt
fastapi==0.115.12
uvicorn==0.34.0
line-bot-sdk==3.13.0
openai==1.75.0
gspread==6.2.1
google-auth==2.39.0
python-dotenv==1.1.0
pydantic==2.11.3
